requirements.txt
tweepy
openai>=1.30.0
python-dotenv
requests
